refactor for easier unit testing
