+++ .gitignore
@@ -1,4 +1,6 @@
 .ipynb_checkpoints
+\#*
+.\#*
 bin
 BIN
 build*
